requirements.txt
fastapi==0.88.0
uvicorn==0.18.3
pydantic==1.10.5
requests==2.28.1
python-dotenv==0.21.0
httpx==0.23.3
